fix: embed works + added transaltions to Embed and Dashboard login
