+++ messages/uk.json
@@ -384,5 +384,25 @@
     "label_city": "Населений пункт:",
     "label_contact": "Контакт:",
     "help_footer": "Якщо вам потрібна допомога із заповненням бланків, зв'яжіться з нами."
+  },
+  "EmbedIndex": {
+    "title": "Віджети для вставки",
+    "description": "Ця сторінка містить віджети для вставки для Радіо TLIS.",
+    "available_widgets": "Доступні віджети:",
+    "show_list_title": "Віджет списку шоу",
+    "show_list_desc": "Відображає список епізодів з плеєром.",
+    "episode_title": "Віджет епізоду",
+    "episode_desc": "Компактний віджет для одного епізоду (квадратний).",
+    "wide_episode_title": "Широкий віджет епізоду",
+    "wide_episode_desc": "Широкий віджет для одного епізоду (горизонтальний)."
+  },
+  "DashboardLogin": {
+    "title": "Вхід до панелі керування",
+    "description": "Увійдіть, щоб отримати доступ до аналітики",
+    "email_label": "Електронна пошта",
+    "password_label": "Пароль",
+    "button": "Увійти",
+    "loading": "Вхід...",
+    "error": "Неправильна пошта або пароль"
   }
 }
